+++ src/ios/CDVFileTransferBCS.m
@@ -18,7 +18,7 @@
  */
 
 #import <Cordova/CDV.h>
-#import "CDVFileTransfer.h"
+#import "CDVFileTransferBCS.h"
 #import "CDVLocalFilesystem.h"
 
 #import <AssetsLibrary/ALAsset.h>
@@ -26,11 +26,11 @@
 #import <AssetsLibrary/ALAssetsLibrary.h>
 #import <CFNetwork/CFNetwork.h>
 
-@interface CDVFileTransfer ()
+@interface CDVFileTransferBCS ()
 // Sets the requests headers for the request.
 - (void)applyRequestHeaders:(NSDictionary*)headers toRequest:(NSMutableURLRequest*)req;
 // Creates a delegate to handle an upload.
-- (CDVFileTransferDelegate*)delegateForUploadCommand:(CDVInvokedUrlCommand*)command;
+- (CDVFileTransferBCSDelegate*)delegateForUploadCommand:(CDVInvokedUrlCommand*)command;
 // Creates an NSData* for the file for the given upload arguments.
 - (void)fileDataForUploadCommand:(CDVInvokedUrlCommand*)command;
 @end
@@ -38,9 +38,9 @@
 // Buffer size to use for streaming uploads.
 static const NSUInteger kStreamBufferSize = 32768;
 // Magic value within the options dict used to set a cookie.
-NSString* const kOptionsKeyCookie = @"__cookie";
+NSString* const kOptionsKeyCookieBCS = @"__cookie";
 // Form boundary for multi-part requests.
-NSString* const kFormBoundary = @"+++++org.apache.cordova.formBoundary";
+// NSString* const kFormBoundary = @"+++++org.apache.cordova.formBoundary";
 
 // Writes the given data to the stream in a blocking way.
 // If successful, returns bytesToWrite.
@@ -69,7 +69,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
     return totalBytesWritten;
 }
 
-@implementation CDVFileTransfer
+@implementation CDVFileTransferBCS
 @synthesize activeTransfers;
 
 - (void)pluginInitialize {
@@ -146,7 +146,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
     // unrecognised values)
     NSString* httpMethod = [command argumentAtIndex:10 withDefault:@"POST"];
     CDVPluginResult* result = nil;
-    CDVFileTransferError errorCode = 0;
+    CDVFileTransferBCSError errorCode = 0;
 
     // NSURL does not accepts URLs with spaces in the path. We escape the path in order
     // to be more lenient.
@@ -170,8 +170,8 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
     [req setHTTPMethod:httpMethod];
 
     //    Magic value to set a cookie
-    if ([options objectForKey:kOptionsKeyCookie]) {
-        [req setValue:[options objectForKey:kOptionsKeyCookie] forHTTPHeaderField:@"Cookie"];
+    if ([options objectForKey:kOptionsKeyCookieBCS]) {
+        [req setValue:[options objectForKey:kOptionsKeyCookieBCS] forHTTPHeaderField:@"Cookie"];
         [req setHTTPShouldHandleCookies:NO];
     }
 
@@ -184,7 +184,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
 
     for (NSString* key in options) {
         id val = [options objectForKey:key];
-        if (!val || (val == [NSNull null]) || [key isEqualToString:kOptionsKeyCookie]) {
+        if (!val || (val == [NSNull null]) || [key isEqualToString:kOptionsKeyCookieBCS]) {
             continue;
         }
         // if it responds to stringValue selector (eg NSNumber) get the NSString
@@ -246,14 +246,14 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
     return req;
 }
 
-- (CDVFileTransferDelegate*)delegateForUploadCommand:(CDVInvokedUrlCommand*)command
+- (CDVFileTransferBCSDelegate*)delegateForUploadCommand:(CDVInvokedUrlCommand*)command
 {
     NSString* source = [command.arguments objectAtIndex:0];
     NSString* server = [command.arguments objectAtIndex:1];
     BOOL trustAllHosts = [[command.arguments objectAtIndex:6 withDefault:[NSNumber numberWithBool:NO]] boolValue]; // allow self-signed certs
     NSString* objectId = [command.arguments objectAtIndex:9];
 
-    CDVFileTransferDelegate* delegate = [[CDVFileTransferDelegate alloc] init];
+    CDVFileTransferBCSDelegate* delegate = [[CDVFileTransferBCSDelegate alloc] init];
 
     delegate.command = self;
     delegate.callbackId = command.callbackId;
@@ -328,7 +328,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
     if (req == nil) {
         return;
     }
-    CDVFileTransferDelegate* delegate = [self delegateForUploadCommand:command];
+    CDVFileTransferBCSDelegate* delegate = [self delegateForUploadCommand:command];
     delegate.connection = [[NSURLConnection alloc] initWithRequest:req delegate:delegate startImmediately:NO];
     if (self.queue == nil) {
         self.queue = [[NSOperationQueue alloc] init];
@@ -351,7 +351,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
     NSString* objectId = [command.arguments objectAtIndex:0];
 
     @synchronized (activeTransfers) {
-        CDVFileTransferDelegate* delegate = activeTransfers[objectId];
+        CDVFileTransferBCSDelegate* delegate = activeTransfers[objectId];
         if (delegate != nil) {
             [delegate cancelTransfer:delegate.connection];
             CDVPluginResult* result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsDictionary:[self createFileTransferError:CONNECTION_ABORTED AndSource:delegate.source AndTarget:delegate.target]];
@@ -370,7 +370,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
     NSDictionary* headers = [command.arguments objectAtIndex:4 withDefault:nil];
 
     CDVPluginResult* result = nil;
-    CDVFileTransferError errorCode = 0;
+    CDVFileTransferBCSError errorCode = 0;
 
     NSURL* targetURL;
 
@@ -406,7 +406,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
     NSMutableURLRequest* req = [NSMutableURLRequest requestWithURL:sourceURL];
     [self applyRequestHeaders:headers toRequest:req];
 
-    CDVFileTransferDelegate* delegate = [[CDVFileTransferDelegate alloc] init];
+    CDVFileTransferBCSDelegate* delegate = [[CDVFileTransferBCSDelegate alloc] init];
     delegate.command = self;
     delegate.direction = CDV_TRANSFER_DOWNLOAD;
     delegate.callbackId = command.callbackId;
@@ -481,7 +481,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
 - (void)onReset {
     @synchronized (activeTransfers) {
         while ([activeTransfers count] > 0) {
-            CDVFileTransferDelegate* delegate = [activeTransfers allValues][0];
+            CDVFileTransferBCSDelegate* delegate = [activeTransfers allValues][0];
             [delegate cancelTransfer:delegate.connection];
         }
     }
@@ -489,18 +489,18 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
 
 @end
 
-@interface CDVFileTransferEntityLengthRequest : NSObject {
+@interface CDVFileTransferBCSEntityLengthRequest : NSObject {
     NSURLConnection* _connection;
-    CDVFileTransferDelegate* __weak _originalDelegate;
+    CDVFileTransferBCSDelegate* __weak _originalDelegate;
 }
 
-- (CDVFileTransferEntityLengthRequest*)initWithOriginalRequest:(NSURLRequest*)originalRequest andDelegate:(CDVFileTransferDelegate*)originalDelegate;
+- (CDVFileTransferBCSEntityLengthRequest*)initWithOriginalRequest:(NSURLRequest*)originalRequest andDelegate:(CDVFileTransferBCSDelegate*)originalDelegate;
 
 @end
 
-@implementation CDVFileTransferEntityLengthRequest
+@implementation CDVFileTransferBCSEntityLengthRequest
 
-- (CDVFileTransferEntityLengthRequest*)initWithOriginalRequest:(NSURLRequest*)originalRequest andDelegate:(CDVFileTransferDelegate*)originalDelegate
+- (CDVFileTransferBCSEntityLengthRequest*)initWithOriginalRequest:(NSURLRequest*)originalRequest andDelegate:(CDVFileTransferBCSDelegate*)originalDelegate
 {
     if (self) {
         DLog(@"Requesting entity length for GZIPped content...");
@@ -529,7 +529,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
 
 @end
 
-@implementation CDVFileTransferDelegate
+@implementation CDVFileTransferBCSDelegate
 
 @synthesize callbackId, connection = _connection, source, target, responseData, responseHeaders, command, bytesTransfered, bytesExpected, direction, responseCode, objectId, targetFileHandle, filePlugin;
 
@@ -594,7 +594,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
 {
     [connection cancel];
     @synchronized (self.command.activeTransfers) {
-        CDVFileTransferDelegate* delegate = self.command.activeTransfers[self.objectId];
+        CDVFileTransferBCSDelegate* delegate = self.command.activeTransfers[self.objectId];
         [self.command.activeTransfers removeObjectForKey:self.objectId];
         [[UIApplication sharedApplication] endBackgroundTask:delegate.backgroundTaskID];
         delegate.backgroundTaskID = UIBackgroundTaskInvalid;
@@ -645,7 +645,7 @@ static CFIndex WriteDataToStream(NSData* data, CFWriteStreamRef stream)
         if ((self.direction == CDV_TRANSFER_DOWNLOAD) && (self.responseCode == 200) && (self.bytesExpected == NSURLResponseUnknownLength)) {
             // Kick off HEAD request to server to get real length
             // bytesExpected will be updated when that response is returned
-            self.entityLengthRequest = [[CDVFileTransferEntityLengthRequest alloc] initWithOriginalRequest:connection.currentRequest andDelegate:self];
+            self.entityLengthRequest = [[CDVFileTransferBCSEntityLengthRequest alloc] initWithOriginalRequest:connection.currentRequest andDelegate:self];
         }
     } else if ([response.URL isFileURL]) {
         NSDictionary* attr = [[NSFileManager defaultManager] attributesOfItemAtPath:[response.URL path] error:nil];