+++ RELEASENOTES.md
@@ -72,7 +72,7 @@
 ### 0.4.3 (Apr 17, 2014)
 * CB-6422 [windows8] use cordova/exec/proxy
 * iOS: Fix error where files were not removed on abort
-* CB-5175: [ios] CDVFileTransfer asynchronous download (Fixes #24)
+* CB-5175: [ios] CDVFileTransferBCS asynchronous download (Fixes #24)
 * [ios] Cast id references to NSURL to avoid compiler warnings (Fixes: apache/cordova-plugin-file-transfer#18)
 * CB-6212: [iOS] fix warnings compiled under arm64 64-bit
 * CB-5762: [FireOS] android: Fix lengthComputable set wrong for gzip downloads